Stop appending the cwd to the file path input to allow absolute paths
The requirements don't state the file path needs to be relative.

The cli documentation was updated to convey any valid file path is acceptable.

The current working directory is no longer appended to the file path input. This was the only use of the os import so it was removed.