Make parse_charger_text_reports throw custom errors
To be more pythonic the parse_charger_text_reports function now catches possible exceptions and raises its own custom errors. Mainly when trying to parse text to numbers from the text_lines.

The points where it would print to console were removed to remove system specific behavior. That should not be a concern of this function.

Also now looks and raises an error for if either section is empty in addition to if they're missing (header line but no data).